009: Faster Rust solution using u32 integers

+++ problems/009/009.cpp
@@ -1,3 +1,4 @@
+#include<cstdint>
 #include<cstring>
 #include<ctime>
 #include<iostream>
@@ -7,16 +8,16 @@ using namespace std;
 
 #define PERIMETER 1000
 
-int solution() {
-  for (int c = PERIMETER / 3 + 1; c < PERIMETER / 2; c++)
-    for (int b = (PERIMETER - c) / 2 + 1; b < c; b++) {
-      int a = PERIMETER - b - c;
+uint64_t solution() {
+  for (uint64_t c = PERIMETER / 3 + 1; c < PERIMETER / 2; c++)
+    for (uint64_t b = (PERIMETER - c) / 2 + 1; b < c; b++) {
+      uint64_t a = PERIMETER - b - c;
 
       if (a * a + b * b == c * c)
         return a * b * c;
     }
 
-  // XXX Don't let the compiler optimize this function away
+  // DON'T OPTIMIZE ME AWAY
   asm("");
 
   return 0;
